// src/utils/interfaces/DocumentVault.interface.ts
export interface IDocument {
  files?: IFile[];
  displayName: string;
  icon: string;
  statusTitle: string;
  status: IDocumentStatus[];
  eVerificationRequired: boolean;
  lastNoOfMonthsRequired: number;
  lastNoOfMonthsGroupBy: "month" | "year";
  endDate: string;
  startDate: string;
  financial: boolean;
  documentNo?: string;
  updatedOn?: string;
  data?: any
}

// ===================== AADHAAR =====================

export interface IAadhaarMethod {
  name: AadhaarVerificationMethods;
}

export interface IAadhaar extends IDocument {
  methods: IAadhaarMethod[];
}
// ...
export interface IPanMethod {
  name: PanVerificationMethods;
}

export interface IPan extends IDocument {
  methods: IPanMethod[];
}

// ===================== GST Certificate =====================

export interface IGstMethod {
  name: GstVerificationMethods;
}

export interface IGST extends IDocument {
  methods: IGstMethod[];
}

// ===================== BANK STATEMENT =====================

export interface IBankStatementMethod {
  name: BankStatementVerificationMethods;
}

export interface IBankStatement extends IDocument {
  methods: IBankStatementMethod[];
}

// ===================== MISCELLANEOUS DOCUMENT =====================

export interface IMiscellaneousDocumentMethod {
  name: MiscellaneousDocumentVerificationMethods;
}

export interface IMiscellaneousDocument extends IDocument {
  methods: IMiscellaneousDocumentMethod[];
}

// ===================== ITR =====================

export interface IItrMethod {
  name: ItrVerificationMethods;
}

export interface IItr extends IDocument {
  methods: IItrMethod[];
}

// ===================== GSTR =====================

export interface IGstrMethod {
  name: GstrVerificationMethods;
}

export interface IGstr extends IDocument {
  methods: IGstrMethod[];
}

// ===================== UDYAM =====================

export interface IUdyamMethod {
  name: UdyamVerificationMethods;
}

export interface IUdyam extends IDocument {
  methods: IUdyamMethod[];
}

// ===================== GSTR FILES =====================

export interface IGstrFileMethod {
  name: GstrFilesVerificationMethods;
}

export interface IGstrFile extends IDocument {
  methods: IGstrFileMethod[];
}

// ===================== CIN =====================

export interface ICINMethod {
  name: CINVerificationMethods;
}

export interface ICIN extends IDocument {
  methods: ICINMethod[];
}

// ===================== CUSTOM_FROM =====================
export interface ICustomFormMethod {
  name: CustomFormVerificationMethods;
}

export interface ICustomForm extends IDocument {
  methods: ICustomFormMethod[];
}

// ===================== PLATFORM DOCUMENTS =====================

export interface IPlatformDocument {
  AADHAAR: IAadhaar;
  PERSONAL_PAN: IPan;
  BUSINESS_PAN: IPan;
  GST: IGST;
  CIN: ICIN;
  BANK_STATEMENT: IBankStatement;
  ITR: IItr;
  GSTR3B_RETSUM: IGstr;
  GSTR2A_B2B: IGstr;
  GSTR2A_RETSUM: IGstr;
  GSTR1_EXP: IGstr;
  GSTR1_B2B: IGstr;
  GSTR1_CDNR: IGstr;
  GSTR1_HSNSUM: IGstr;
  GSTR1_CDNUR: IGstr;
  GSTR1_B2CL: IGstr;
  GSTR1_CDN: IGstr;
  GSTR1_NIL: IGstr;
  GSTR1_B2CS: IGstr;
  UDYAM: IUdyam;
  GSTR_FILES: IGstrFile;
  MISCELLANEOUS_DOCUMENT: IMiscellaneousDocument;
  CUSTOM_FROM: ICustomForm;
}
// ...
export interface IStep {
  title: string;
  description: string;
  status: "error" | "finish" | "process" | "wait";
}